requirements.txt
pandas==2.1.4
numpy==1.24.3
scikit-learn==1.3.2
ccxt==4.1.77
ta==0.10.2
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.17.0
python-dotenv==1.0.0
requests==2.31.0
yfinance==0.2.28
python-telegram-bot==20.7 